django/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-environ
shortuuid
django-jazzmin
python-dotenv
Pillow
